+++ canopy/study_result.py
@@ -3,12 +3,14 @@ from typing import Sequence, Union, Optional
 import canopy
 
 
-class StudyDataResult:
+class StudyResult:
+    _data: Optional[canopy.swagger.StudyDocument] = None
+
     def __init__(
             self,
             session: canopy.Session,
             study_result: Optional[canopy.swagger.GetStudyQueryResult],
-            jobs: Sequence[canopy.StudyJobDataResult],
+            jobs: Sequence[canopy.StudyJobResult],
             document: Optional[canopy.swagger.CanopyDocument] = None):
         self._session = session
         self._study_result = study_result
@@ -16,7 +18,7 @@ class StudyDataResult:
         self._jobs = jobs
 
     @property
-    def study_result(self) -> canopy.swagger.GetStudyQueryResult:
+    def result(self) -> canopy.swagger.GetStudyQueryResult:
         return self._study_result
 
     @property
@@ -24,9 +26,11 @@ class StudyDataResult:
         return self._document
 
     @property
-    def study_document(self) -> canopy.swagger.StudyDocument:
-        return canopy.get_study_document(self._session, self._document)
+    def data(self) -> canopy.swagger.StudyDocument:
+        if self._data is None:
+            self._data = canopy.get_study_document(self._session, self._document)
+        return self._data
 
     @property
-    def jobs(self) -> Sequence[canopy.StudyJobDataResult]:
+    def jobs(self) -> Sequence[canopy.StudyJobResult]:
         return self._jobs